Rename node langNames to localeDisplayNames
